+++ src/services/auth/index.ts
@@ -1,56 +1,34 @@
-import jwt, { Secret, SignOptions, VerifyOptions } from 'jsonwebtoken';
-import { promisify } from 'util';
-
 import fastifyAuth from '@fastify/auth';
 import fastifyBearerAuth from '@fastify/bearer-auth';
 import fastifyCors from '@fastify/cors';
 import fastifySecureSession from '@fastify/secure-session';
-import { FastifyPluginAsync, FastifyRequest } from 'fastify';
-
-import { MAIL } from '@graasp/translations';
+import { FastifyPluginAsync } from 'fastify';
 
 import {
-  AUTH_TOKEN_EXPIRATION_IN_MINUTES,
-  AUTH_TOKEN_JWT_SECRET,
-  JWT_SECRET,
-  LOGIN_TOKEN_EXPIRATION_IN_MINUTES,
-  MOBILE_AUTH_URL,
   PROD,
-  PUBLIC_URL,
   RECAPTCHA_SECRET_ACCESS_KEY,
-  REFRESH_TOKEN_EXPIRATION_IN_MINUTES,
-  REFRESH_TOKEN_JWT_SECRET,
-  REGISTER_TOKEN_EXPIRATION_IN_MINUTES,
   SECURE_SESSION_SECRET_KEY,
   STAGING,
   TOKEN_BASED_AUTH,
 } from '../../utils/config';
-import { InvalidSession, OrphanSession } from '../../utils/errors';
-import { Member } from '../member/entities/member';
-import MemberRepository from '../member/repository';
 import { AuthPluginOptions } from './interfaces/auth';
 import captchaPlugin from './plugins/captcha';
 import magicLinkController from './plugins/magicLink';
 import mobileController from './plugins/mobile';
 import passwordController from './plugins/password';
-import { getRedirectionUrl } from './utils';
-
-const promisifiedJwtVerify = promisify<
-  string,
-  Secret,
-  VerifyOptions,
-  { sub: string; challenge?: string }
->(jwt.verify);
-const promisifiedJwtSign = promisify<
-  { sub: string; challenge?: string },
-  Secret,
-  SignOptions,
-  string
->(jwt.sign);
+import { AuthService } from './service';
+import {
+  fetchMemberInSession,
+  generateAuthTokensPair,
+  verifyMemberInAuthToken,
+  verifyMemberInSession,
+} from './utils';
 
 const plugin: FastifyPluginAsync<AuthPluginOptions> = async (fastify, options) => {
   const { sessionCookieDomain: domain } = options;
   const { log, mailer } = fastify;
+
+  const authService = new AuthService(mailer, log);
 
   // cookie based auth
   await fastify.register(fastifySecureSession, {
@@ -61,67 +39,9 @@ const plugin: FastifyPluginAsync<AuthPluginOptions> = async (fastify, options) =
   // captcha
   await fastify.register(captchaPlugin, { secretAccessKey: RECAPTCHA_SECRET_ACCESS_KEY });
 
-  async function verifyMemberInSession(request: FastifyRequest) {
-    const { session } = request;
-    const memberId = session.get('member');
-
-    if (!memberId) {
-      throw new InvalidSession();
-    }
-
-    // TODO: do we really need to get the user from the DB? (or actor: { id } is enough?)
-    // maybe when the groups are implemented it will be necessary.
-    const member = await MemberRepository.findOneBy({ id: memberId });
-
-    if (!member) {
-      session.delete();
-      throw new OrphanSession();
-    }
-
-    request.member = member;
-  }
-
-  // set member in request from session if exist
-  // used to get authenticated member without throwing
-  async function fetchMemberInSession(request: FastifyRequest) {
-    const { session } = request;
-    const memberId = session.get('member');
-
-    if (!memberId) return;
-
-    // TODO: do we really need to get the user from the DB? (or actor: { id } is enough?)
-    // maybe when the groups are implemented it will be necessary.
-    request.member = (await MemberRepository.findOneBy({ id: memberId })) ?? undefined;
-  }
-
   fastify.decorate('fetchMemberInSession', fetchMemberInSession);
 
   fastify.decorate('validateSession', verifyMemberInSession);
-
-  // for token based auth
-  async function verifyMemberInAuthToken(jwtToken: string, request: FastifyRequest) {
-    try {
-      const { routerPath } = request;
-      const refreshing = '/m/auth/refresh' === routerPath;
-      const secret = refreshing ? REFRESH_TOKEN_JWT_SECRET : AUTH_TOKEN_JWT_SECRET;
-
-      const { sub: memberId } = await promisifiedJwtVerify(jwtToken, secret, {});
-
-      if (refreshing) {
-        request.memberId = memberId;
-      } else {
-        const member = await MemberRepository.findOneBy({ id: memberId });
-        if (!member) return false;
-        request.member = member;
-      }
-
-      return true;
-    } catch (error) {
-      const { log } = request;
-      log.warn('Invalid auth token');
-      return false;
-    }
-  }
 
   await fastify.register(fastifyAuth);
   await fastify.register(fastifyBearerAuth, {
@@ -131,7 +51,7 @@ const plugin: FastifyPluginAsync<AuthPluginOptions> = async (fastify, options) =
   });
 
   if (!fastify.verifyBearerAuth) {
-    throw new Error();
+    throw new Error('verifyBearerAuth is not defined');
   }
 
   fastify.decorate(
@@ -154,93 +74,12 @@ const plugin: FastifyPluginAsync<AuthPluginOptions> = async (fastify, options) =
 
   fastify.decorate('verifyAuthentication', verifyAuthentication);
 
-  const getLangFromMember = (member, defaultLang?): string => member.extra?.lang ?? defaultLang;
-
-  async function generateAuthTokensPair(
-    memberId: string,
-  ): Promise<{ authToken: string; refreshToken: string }> {
-    const [authToken, refreshToken] = await Promise.all([
-      promisifiedJwtSign({ sub: memberId }, AUTH_TOKEN_JWT_SECRET, {
-        expiresIn: `${AUTH_TOKEN_EXPIRATION_IN_MINUTES}m`,
-      }),
-      promisifiedJwtSign({ sub: memberId }, REFRESH_TOKEN_JWT_SECRET, {
-        expiresIn: `${REFRESH_TOKEN_EXPIRATION_IN_MINUTES}m`,
-      }),
-    ]);
-    return { authToken, refreshToken };
-  }
   fastify.decorate('generateAuthTokensPair', generateAuthTokensPair);
 
-  async function generateRegisterLinkAndEmailIt(
-    member,
-    options: { challenge?; url?: string } = {},
-  ) {
-    const { challenge, url } = options;
+  // TODO: decorate auth service and use it instead of decorating function
+  fastify.decorate('generateRegisterLinkAndEmailIt', authService.generateRegisterLinkAndEmailIt);
 
-    // generate token with member info and expiration
-    const token = await promisifiedJwtSign({ sub: member.id, challenge }, JWT_SECRET, {
-      expiresIn: `${REGISTER_TOKEN_EXPIRATION_IN_MINUTES}m`,
-    });
-
-    const redirectionUrl = getRedirectionUrl(url);
-    const domain = challenge ? MOBILE_AUTH_URL : PUBLIC_URL;
-    const destination = new URL('/auth', domain);
-    destination.searchParams.set('t', token);
-    destination.searchParams.set('url', redirectionUrl);
-    const link = destination.toString();
-
-    const lang = getLangFromMember(member);
-
-    const translated = mailer.translate(lang);
-    const subject = translated(MAIL.SIGN_UP_TITLE);
-    const html = `
-    ${mailer.buildText(translated(MAIL.GREETINGS))}
-    ${mailer.buildText(translated(MAIL.SIGN_UP_TEXT))}
-    ${mailer.buildButton(link, translated(MAIL.SIGN_UP_BUTTON_TEXT))}
-    ${mailer.buildText(translated(MAIL.SIGN_UP_NOT_REQUESTED))}`;
-
-    // don't wait for mailer's response; log error and link if it fails.
-    mailer
-      .sendEmail(subject, member.email, link, html)
-      .catch((err) => log.warn(err, `mailer failed. link: ${link}`));
-  }
-
-  fastify.decorate('generateRegisterLinkAndEmailIt', generateRegisterLinkAndEmailIt);
-
-  async function generateLoginLinkAndEmailIt(
-    member: Member,
-    options: { challenge?: string; lang?: string; url?: string } = {},
-  ) {
-    const { challenge, lang, url } = options;
-
-    // generate token with member info and expiration
-    const token = await promisifiedJwtSign({ sub: member.id, challenge }, JWT_SECRET, {
-      expiresIn: `${LOGIN_TOKEN_EXPIRATION_IN_MINUTES}m`,
-    });
-
-    const redirectionUrl = getRedirectionUrl(url);
-    const domain = challenge ? MOBILE_AUTH_URL : PUBLIC_URL;
-    const destination = new URL('/auth', domain);
-    destination.searchParams.set('t', token);
-    destination.searchParams.set('url', redirectionUrl);
-    const link = destination.toString();
-
-    const memberLang = getLangFromMember(member) ?? lang;
-
-    const translated = mailer.translate(memberLang);
-    const subject = translated(MAIL.SIGN_IN_TITLE);
-    const html = `
-    ${mailer.buildText(translated(MAIL.SIGN_IN_TEXT))}
-    ${mailer.buildButton(link, translated(MAIL.SIGN_IN_BUTTON_TEXT))}
-    ${mailer.buildText(translated(MAIL.SIGN_IN_NOT_REQUESTED))}`;
-
-    // don't wait for mailer's response; log error and link if it fails.
-    fastify.mailer
-      .sendEmail(subject, member.email, link, html)
-      .catch((err) => log.warn(err, `mailer failed. link: ${link}`));
-  }
-
-  fastify.decorate('generateLoginLinkAndEmailIt', generateLoginLinkAndEmailIt);
+  fastify.decorate('generateLoginLinkAndEmailIt', authService.generateLoginLinkAndEmailIt);
 
   fastify.register(async function (fastify) {
     // add CORS support